### 2024-09-12 #7219 @crvernon
👋 @jorisvandenbossche - Would you be willing to review this submission to JOSS?  Should be right down your alley.  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
