### 2024-09-16 #7219 @hariharanragothaman
@crvernon Thanks for your response. Appreciate it. 🫡 👍🏻

1. To give some background I was looking for reviewing opportunities and I had signed up to be a reviewer.
2. My JOSS reviewer profile for reference: https://reviewers.joss.theoj.org/reviewers/4516
3. I filtered through issues in this repository - specifically on Track5 & Track7 - which falls under my areas of interest and expertise.

Happy to provide more details and context.
