### 2024-09-16 #7239 @editorialbot
[REVIEW]: xagg: A Python package to aggregate gridded data onto polygons
**Submitting author:** @ks905383 (<a href="http://orcid.org/0000-0001-8309-7124">Kevin Schwarzwald</a>)
**Repository:** https://github.com/ks905383/xagg/
**Branch with paper.md** (empty if default branch): joss_submission
**Version:**  v0.3.2.4
