### 2024-09-24 734a372
Fix for 4d aggregate
