### 2024-09-27 27e90eb
Add partial nan error to dot product pathway
- Makes sure partial nan warning is raised on dot product implementation
- Makes partial nan warning a little clearer
- Makes sure to only trigger it once per code run, instead of for every poly_idx / variable
