### 2024-09-27 46d777a
fix test error with extra `copy.deepcopy()` calls
