### 2024-09-27 a74111b
remove joss paper from this branch
