### 2024-10-01 78864ce
Add CITATION.cff file
Closes #82
