### 2024-10-01 c4af3de
Update CITATION.cff
