### 2024-10-01 89eaeec
Update CITATION.cff
