### 2024-10-01 b2ab876
Revert "Merge pull request #83 from ks905383/fix_dotprod_extradims"
This reverts commit 787299d95271c726b0b78318c6dfe423d14ed2c7, reversing
changes made to c46a9247b8d617abc53a97ca98b4142813c9aa8f.
