+++ tests/test_core.py
@@ -1,11 +1,9 @@
 import pytest
-import copy
 import pandas as pd
 import numpy as np
 import xarray as xr
 import geopandas as gpd
 from geopandas import testing as gpdt
-import itertools
 from unittest import TestCase
 from shapely.geometry import Polygon
 from unittest.mock import patch
@@ -203,14 +201,13 @@ def test_create_raster_polygons_at180():
 
 
 ##### get_pixel_overlaps() tests #####
-# build raster polygons from a simple 2x2x3 grid of lat/lon/time pixels
-ds = xr.Dataset({'test':(['lon','lat','time'],np.reshape(np.arange(1,13),(2,2,3))),
+# build raster polygons from a simple 2x2 grid of lat/lon pixels
+ds = xr.Dataset({'test':(['lon','lat'],np.array([[0,1],[2,3]])),
 				 'lat_bnds':(['lat','bnds'],np.array([[-0.5,0.5],[0.5,1.5]])),
 				 'lon_bnds':(['lon','bnds'],np.array([[-0.5,0.5],[0.5,1.5]]))},
 				coords={'lat':(['lat'],np.array([0,1])),
 						'lon':(['lon'],np.array([0,1])),
-						'bnds':(['bnds'],np.array([0,1])),
-						'time':(['time'],pd.date_range('2019-01-01','2019-01-03'))})
+						'bnds':(['bnds'],np.array([0,1]))})
 
 # add a simple weights grid
 weights = xr.DataArray(data=np.array([[0,1],[2,3]]),
@@ -224,7 +221,7 @@ pix_agg = create_raster_polygons(ds,weights=weights)
 
 
 # Test if the shapefile completely covers one pixel
-def test_get_pixel_overlaps_one_pixel(pix_agg=copy.deepcopy(pix_agg)):
+def test_get_pixel_overlaps_one_pixel(pix_agg=pix_agg):
 	# Create polygon covering one pixel
 	gdf_test = {'name':['test'],
 				'geometry':[Polygon([(-0.5,-0.5),(-0.5,0.5),(0.5,0.5),(0.5,-0.5),(-0.5,-0.5)])]}
@@ -252,7 +249,7 @@ def test_get_pixel_overlaps_one_pixel(pix_agg=copy.deepcopy(pix_agg)):
 
 
 # Test if the shapefile is a fraction of one pixel
-def test_get_pixel_overlaps_fraction_of_pixel(pix_agg=copy.deepcopy(pix_agg)):
+def test_get_pixel_overlaps_fraction_of_pixel(pix_agg=pix_agg):
 	# Create polygon covering less than one pixel
 	gdf_test = {'name':['test'],
 				'geometry':[Polygon([(-0.5,-0.5),(-0.5,0),(0,0),(0,-0.5),(-0.5,-0.5)])]}
@@ -279,7 +276,7 @@ def test_get_pixel_overlaps_fraction_of_pixel(pix_agg=copy.deepcopy(pix_agg)):
 	assert np.allclose([v for v in df0.coords],[v for v in df_compare.coords])
 
 # Test if the shapefile perfectly covers several complete pixels
-def test_get_pixel_overlaps_multiple_pixels_complete(pix_agg=copy.deepcopy(pix_agg)):
+def test_get_pixel_overlaps_multiple_pixels_complete(pix_agg=pix_agg):
 	# Create polygon covering multiple pixels
 	gdf_test = {'name':['test'],
 				'geometry':[Polygon([(-0.5,-0.5),(-0.5,1.5),(1.5,1.5),(1.5,-0.5),(-0.5,-0.5)])]}
@@ -309,7 +306,7 @@ def test_get_pixel_overlaps_multiple_pixels_complete(pix_agg=copy.deepcopy(pix_a
 
 
[…]
