### 2024-10-01 49231a1
Reapply "Merge pull request #83 from ks905383/fix_dotprod_extradims"
This reverts commit b2ab8760c32ad44d4ad637a6dcd9a833492fd34f.
