### 2024-10-01 2a98077
Revert "Update CITATION.cff"
This reverts commit 89eaeeca7cbc40f3ac7456fea5b43a4333466bbf.
