### 2024-10-01 842d061
Revert "Update CITATION.cff"
This reverts commit c4af3de329ada5f9589739fa72c3952c5616c4f5.
