### 2024-10-01 da075aa
Revert "Add CITATION.cff file"
This reverts commit 78864ce7e1de3b2c885b554d97022145b4ab5f47.
