### 2024-10-01 cd09304
Create CITATION.cff
