+++ README.md
@@ -15,10 +15,10 @@ A package to aggregate gridded data in `xarray` to polygons in `geopandas` using
 The easiest way to install the latest version of `xagg` is using `conda` or `mamba`:  
 
 ```
-conda install -c conda-forge xagg==0.3.2.3
+conda install -c conda-forge xagg==0.3.2.4
 
 # or
-mamba install -c conda-forge xagg==0.3.2.3
+mamba install -c conda-forge xagg==0.3.2.4
 ```
 
 We recommend installing `xagg` in a new environment whenever possible, to ensure all (sub)dependencies are correctly loaded. 
