### 2024-10-25 #7239 @thurber
hi @crvernon, my review is in process, i logged a couple minor issues today; once those are resolved i can wrap up.

@ks905383, the review checklist includes confirming that a community/contribution guideline exists as part of your documentation. Can you add something like that, or point me toward it if i've missed it?
