### 2024-10-28 #7239 @ks905383
> @ks905383, the review checklist includes confirming that a community/contribution guideline exists as part of your documentation. Can you add something like that, or point me toward it if i've missed it?

Ah yes, that's definitely missing right now, thanks for pointing that out. Do you have any favorite templates? Otherwise I'll look around for a good example to work off of.
