### 2024-10-28 #7239 @thurber
> Ah yes, that's definitely missing right now, thanks for pointing that out. Do you have any favorite templates? Otherwise I'll look around for a good example to work off of.

The GitHub guidelines are pretty good and point to some examples: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
