### 2024-10-28 ks905383/xagg#86 @thurber
[xagg_joss_review.zip](https://github.com/user-attachments/files/17547760/xagg_joss_review.zip)

Here's a sample with a netcdf with a `band_data` temperature variable and a random `weight` variable, and a geojson with counties in WA state.
