### 2024-10-29 cb0b062
Address bbox subsetting weights bug
