+++ xagg/core.py
@@ -174,6 +174,21 @@ def process_weights(ds,weights=None,target='ds',silent=None):
                                  '`xesmf` is needed for `xagg` to regrid them to match; however, '+
                                  '`xesmf` is not installed. Either install `xesmf` or '+
                                  'manually regrid them to match each other.')
+
+            # Make sure the weights actually cover the dataset before 
+            # regridding (which could otherwise interpolate / hallucinate
+            # beyond its boundaries)
+            bbox_ds = [ds.lat.min(),ds.lat.max(),ds.lon.min(),ds.lon.max()]
+            bbox_wgts = [weights.lat.min(),weights.lat.max(),weights.lon.min(),weights.lon.max()]
+
+            if ((bbox_wgts[0]>bbox_ds[0]) or (bbox_wgts[1]<bbox_ds[1]) or
+                (bbox_wgts[2]>bbox_ds[2]) or (bbox_wgts[3]<bbox_wgts[3])):
+                warnings.warn('The `weights` input (bbox latmin, latmax, lonmin, lonmax: '+
+                              ', '.join([str(k).values for k in bbox_wgts])+
+                              ' spans less area than the `ds` input (bbox: '+
+                              ', '.join([str(k).values for k in bbox_ds])+'). Weights beyond '+
+                              ' this bounding box may be set to 0 or incorrectly interpolated.')
+
             if target == 'ds':
                 if not silent:
                     print('regridding weights to data grid...')
@@ -293,6 +308,10 @@ def create_raster_polygons(ds,
     ds = fix_ds(ds)
     ds = get_bnds(ds,silent=silent)
     #breakpoint()
+    
+    # Process weights
+    ds,winf = process_weights(ds,weights,target=weights_target)
+
     # Subset by shapefile bounding box, if desired
     if subset_bbox is not None:
         if type(subset_bbox) == gpd.geodataframe.GeoDataFrame:
@@ -303,11 +322,13 @@ def create_raster_polygons(ds,
             bbox_thresh = grid_dist*2. # then set threshold to twice grid size, avoids huge subsets for high res grids
             ds = ds.sel(lon=slice(subset_bbox.total_bounds[0]-bbox_thresh,subset_bbox.total_bounds[2]+bbox_thresh),
                         lat=slice(subset_bbox.total_bounds[1]-bbox_thresh,subset_bbox.total_bounds[3]+bbox_thresh))
+            if weights is not None:
+                # Also subset weights (which should now match the ds grid after `process_weights`)
+                weights = weights.sel(lon=slice(subset_bbox.total_bounds[0]-bbox_thresh,subset_bbox.total_bounds[2]+bbox_thresh),
+                                  lat=slice(subset_bbox.total_bounds[1]-bbox_thresh,subset_bbox.total_bounds[3]+bbox_thresh))
         else:
             warnings.warn('[subset_bbox] is not a geodataframe; no mask by polygon bounding box used.')
             
-    # Process weights
-    ds,winf = process_weights(ds,weights,target=weights_target)
             
     # Mask
     if mask is not None:
