### 2024-10-29 44d9c6f
make sure weights subset warning is in the right place
