+++ tests/test_core.py
@@ -1,4 +1,5 @@
 import pytest
+import warnings
 import copy
 import pandas as pd
 import numpy as np
