### 2024-10-29 e979fe8
Fix groupby deprecation warnings in dot product calc
