### 2024-10-29 31d91d9
Implement non-deprecated pandas groupby in dotprod impl
Using https://stackoverflow.com/questions/78542997/deprecation-warning-with-groupby-apply-problem-using-with-lambda-and-group/78543185#78543185 to keep the grouping column and silence the warning.
