### 2024-10-29 2ebbd06
remove weights bbox subsetting since already captured by ds
