### 2024-10-29 06c9bbc
finish writing test for #86
