### 2024-11-01 5d51efa
Add contribution guidelines
