### 2024-11-01 333fa9a
update docs with contributing guidelines and IndexError issue
