+++ docs/source/tips.rst
@@ -1,4 +1,4 @@
-Tips 
+Tips
 #######################################
 
 Silencing status updates
@@ -61,6 +61,9 @@ Once you have created a :py:class:`weightmap`, the following code will create a 
    # Plotting the first polygon in the polygon `gdf`
    weightmap.diag_fig(0,ds)
 
+IndexErrors in :py:meth:`xa.pixel_overlaps`
+------------------------------------------------------------
+If you're running into an `IndexError` when running :py:meth:`xa.pixel_overlaps` (e.g., `IndexError: too many indices for array: array is 1-dimensional, but 3 were indexed`), try reprojecting the input :py:meth:`geodataframe` to `'EPSG:4326'` before running :py:meth:`xa.pixel_overlaps`. See `here <https://github.com/ks905383/xagg/issues/80>`_ for more discussion on this issue.
 
 
 
