### 2024-11-01 #7239 @ks905383
Thanks @thurber for the raised issues and suggestions! I've fixed the bugs that you ran into (the fixes are in the main branch; I'll include them in the next update with any other fixes from this review, unless you'd like to check them earlier) and added contribution guidelines to the README and the docs.
