### 2024-11-01 216f71d
fix bulleted list in the docs
