### 2024-11-06 #7239 @thurber
@ks905383, I've tested the updates you made to address my issues, and all looks good! I have a couple final thoughts below, but otherwise feel the publication is ready. Thanks for this useful software!

- [ ] In your updates last week, a couple files got added to the root of the repo that probably don't belong: one file just named `=` and another named `=0.7.1`.
- [ ] I think it is important to make it more obvious in both the paper and the docs that the software is currently only designed to work with data and shapes defined on an equirectangular projection, and additionally that the data must be defined on a rectangular grid. It would be amazing to eventually support other projections and non-rectangular grids, but I recognize that as being difficult and out of scope.
