### 2024-11-12 47de3ed
add mention of rectangular grid / proj req. to docs
