### 2024-11-12 297aecd
add reference to rectangular grid requirement
