+++ joss_paper/paper.md
@@ -35,7 +35,7 @@ When these raster and polygon worlds collide, as they often do in social or natu
 
 Therefore, an ideal aggregation would weight not only by the area of overlap between grid cells and polygons, but also optionally by other densities of relevant variables - population, area planted, etc. [@auffhammer_using_2013].
 
-`xagg` fulfills this need, by providing a simple interface for aggregating raster data stored in `xarray` [@hoyer_xarray_2017] `Datasets` or `DataArrays` onto polygons stored in `geopandas` [@bossche_geopandasgeopandas_2024] `geodataframes`, weighted by the fractional area overlap between the raster grid and the polygon, and optionally additionally weighted by a secondary gridded variable (see \autoref{fig1} for a sample workflow). Fractional area weights are generated by constructing polygons for each grid cell and using `geopandas`' `gpd.overlay()` function to calculate the overlaps between input polygons and grid cells. Aggregated data is then returned as an `xarray` `Dataset`, a `pandas` `DataFrame`, or a `geopandas` `GeoDataFrame`, depending on the user's needs.  
+`xagg` fulfills this need, by providing a simple interface for aggregating raster data stored in rectangular grids in `xarray` [@hoyer_xarray_2017] `Datasets` or `DataArrays` onto polygons stored in `geopandas` [@bossche_geopandasgeopandas_2024] `geodataframes`, weighted by the fractional area overlap between the raster grid and the polygon, and optionally additionally weighted by a secondary gridded variable (see \autoref{fig1} for a sample workflow). Fractional area weights are generated by constructing polygons for each grid cell and using `geopandas`' `gpd.overlay()` function to calculate the overlaps between input polygons and grid cells. Aggregated data is then returned as an `xarray` `Dataset`, a `pandas` `DataFrame`, or a `geopandas` `GeoDataFrame`, depending on the user's needs.  
 
 
 # Statement of need
