### 2024-11-12 5e38b3b
doc tweak on rectangular grids
