### 2024-11-12 #7239 @ks905383
@thurber great, thanks!

I've removed the empty files (I think `xesmf`-caused at some point) and tweaked the language in the paper. Strictly speaking, the package's only requirement on projection of the shapes is that it is re-projectable to the EASE grids (`EPSG:6931-6933`) by `geopandas` itself; but there can be a confusing error when that isn't the case. I've updated the docs [here](https://xagg.readthedocs.io/en/latest/tips.html#indexerrors-in-xa-pixel-overlaps) to mention this issue.
