+++ .github/workflows/test.yaml
@@ -31,7 +31,7 @@ jobs:
         run: |
           pytest --cov=./ --cov-report=xml
       - name: Upload code coverage to Codecov
-        uses: codecov/codecov-action@v4
+        uses: codecov/codecov-action@v5
         with:
           file: ./coverage.xml
           token: ${{ secrets.CODECOV_TOKEN }} 
