### 2024-12-16 #7239 @hariharanragothaman
1. I have reviewed completely. Apologies for the latest update, thanks a lot for your patience. (Had a severe eye infection)
2. There is only one broken link in the read-me in github repo, reference: https://github.com/ks905383/xagg/blob/main/ko-fi.com/ks905383 @ks905383 - Feel free to fix it when you have time.

4. Otherwise, we are all set!  I have also updated my review check-list
Thanks a lot for the opportunity! 😃  ✅
