### 2024-12-24 #7239 @crvernon
:wave: @ks905383 - I just have the following few things I need you to address before I move to accept this for publication:

In the archive:

- [ ] I listed the current version of the software as v0.3.2.4, let me know if this is not correct.
- [ ] Your Zenodo archive has a different license then your GitHub repo.  Please edit the metadata in your Zenodo archive to match.

In the paper:

- [ ] LINE 52:  Not sure if there is supposed to be an * in front of "pandas"
- [ ] LINE 56:  The "p" in "python" should be capitalized
- [ ] LINE 84:  No need for the * at the end of the publication title.  I believe this was a journal specific inclusion, not actually part of the title of the work.

Let me know once these have been addressed.  Thanks!
