+++ joss_paper/paper.md
@@ -41,9 +41,9 @@ Therefore, an ideal aggregation would weight not only by the area of overlap bet
 # Statement of need
 Aggregating gridded data onto polygons is a fundamental aspect of much social and natural science research (e.g., @auffhammer_using_2013; @hsiang_estimating_2017; @carleton_valuing_2022; @mastrantonas_forecasting_2022). Historically, this process has been conducted on an ad hoc basis by individual research groups, often using simplifications such as averaging over all grid cells that overlap with a county, regardless of the size of that overlap (e.g,. @schlenker_nonlinear_2009). 
 
-`xagg` fills a need for an easy, standardized, and accurate workflow for this aggregation. Accepting and outputting data in `xarray` and `*pandas` formats (including keeping by default relevant metadata and attributes from the inputted polygons) means `xagg` can be plugged into a wide array of existing workflows in natural and social sciences, and can easily export aggregated results in formats read by other languages often used in research, including R, QGIS, or STATA. 
+`xagg` fills a need for an easy, standardized, and accurate workflow for this aggregation. Accepting and outputting data in `xarray` and `pandas`/`geopandas` formats (including keeping by default relevant metadata and attributes from the inputted polygons) means `xagg` can be plugged into a wide array of existing workflows in natural and social sciences, and can easily export aggregated results in formats read by other languages often used in research, including R, QGIS, or STATA. 
 
-Though other `python` packages facilitate the aggregation of raster data, to the authors' knowledge, none provide the same depth of functionality or conduct the final aggregation internally. The `mask_3D_frac_approx` function from the `regionmask` package [@hauser_regionmaskregionmask_2023] also creates weights from relative overlaps between grid cells and regions, for example; this however only works for regular rectangular grids (while `xagg` works with any rectangular grid), and results in more approximate overlaps than those calculated using `xagg`. In addition, none allow easy weighting by a secondary raster variable (e.g., population density or yield), or keep polygon metadata intact (which is often needed to merge in other datasets after aggregation). 
+Though other Python packages facilitate the aggregation of raster data, to the authors' knowledge, none provide the same depth of functionality or conduct the final aggregation internally. The `mask_3D_frac_approx` function from the `regionmask` package [@hauser_regionmaskregionmask_2023] also creates weights from relative overlaps between grid cells and regions, for example; this however only works for regular rectangular grids (while `xagg` works with any rectangular grid), and results in more approximate overlaps than those calculated using `xagg`. In addition, none allow easy weighting by a secondary raster variable (e.g., population density or yield), or keep polygon metadata intact (which is often needed to merge in other datasets after aggregation). 
 
 `xagg` has already been used in peer-reviewed (e.g., @pulla_grace_2023-1; @mastrantonas_forecasting_2022; @schwarzwald_importance_2022) and upcoming (e.g., @sichone_assessment_2024; @peard_combining_2023) scientific publications, has reached over 15,000 cumulative downloads across versions, and is a key component of a how-to guide for climate econometrics [@rising_practical_2024]. 
 
