+++ joss_paper/paper.bib
@@ -105,7 +105,7 @@
 }
 
 @article{carleton_valuing_2022,
-  title = {Valuing the {{Global Mortality Consequences}} of {{Climate Change Accounting}} for {{Adaptation Costs}} and {{Benefits}}*},
+  title = {Valuing the {{Global Mortality Consequences}} of {{Climate Change Accounting}} for {{Adaptation Costs}} and {{Benefits}}},
   author = {Carleton, Tamma and Jina, Amir and Delgado, Michael and Greenstone, Michael and Houser, Trevor and Hsiang, Solomon and Hultgren, Andrew and Kopp, Robert E and McCusker, Kelly E and Nath, Ishan and Rising, James and Rode, Ashwin and Seo, Hee Kwon and Viaene, Arvid and Yuan, Jiacan and Zhang, Alice Tianbo},
   year = {2022},
   month = nov,
