### 2024-12-31 #7239 @ks905383
@crvernon sounds great! `v0.3.2.4` is indeed the current version. I've fixed the other issues you've raised as well.
